\name{gplots-deprecated}
\alias{deprecated}
\title{Deprecated Functions}
\description{
  These functions have been deprecated and will be removed in future
  releases of gplots.
}
\details{
  These functions have been deprecated.  Please refer to the manual page
  for the replacement function:
  \itemize{
    \item (No deprecated functions at this time)
  }
}
\author{Gregory R. Warnes \email{greg@warnes.net}}
\seealso{
  \code{\link[base]{Deprecated}}.

  \code{\link{gplots-package}} gives an overview of the package.
}
\keyword{misc}
