\name{residplot}
\alias{residplot}
\title{Undocumented Functions}
\description{
  These functions are undocumented. Some are internal and not intended
  for direct use. Others simply haven't been documented yet.
}
\usage{
residplot(model, formula, \dots)
}
\arguments{
  \item{model}{undocumented.}
  \item{formula}{undocumented.}
  \item{\dots}{undocumented.}
}
\author{Gregory R. Warnes \email{greg@warnes.net}.}
\seealso{
  \code{\link{gplots-package}} gives an overview of the package.
}
\keyword{misc}
