% $Id: residplot.Rd 1319 2009-05-08 22:01:12Z warnes $

\name{residplot}
\alias{residplot}
%\alias{iapply}
\title{Undocumented functions}
\description{
  These functions are undocumented.  Some are internal and not intended
  for direct use.  Others simply haven't been documented yet.  
}
\usage{
residplot(model, formula, ...) 
}
\arguments{
  \item{model}{Undocumented}
  \item{formula}{Undocumented}
    \item{\dots}{ arguments to be passed to \code{fun} }
  }
\details{
 These functions are undocumented.  Some are internal and not intended
   for direct use.  Others simply haven't been documented yet.
   }
\author{Gregory R. Warnes \email{greg@random-technologies-llc.com}}
\keyword{misc}
