% $Id: residplot.Rd,v 1.2 2005/06/09 14:20:28 nj7w Exp $

\name{residplot}
\alias{residplot}
%\alias{iapply}
\title{Undocumented functions}
\description{
  These functions are undocumented.  Some are internal and not intended
  for direct use.  Others simply haven't been documented yet.  
}
\usage{
residplot(model, formula, ...) 
}
\arguments{
  \item{model}{Undocumented}
  \item{formula}{Undocumented}
    \item{\dots}{ arguments to be passed to \code{fun} }
  }
\details{
 These functions are undocumented.  Some are internal and not intended
   for direct use.  Others simply haven't been documented yet.
   }
\author{Gregory R. Warnes \email{gregory\_r\_warnes@groton.pfizer.com}}
\keyword{misc}
