% $Id: residplot.Rd 671 2005-09-12 15:44:37Z nj7w $

\name{residplot}
\alias{residplot}
%\alias{iapply}
\title{Undocumented functions}
\description{
  These functions are undocumented.  Some are internal and not intended
  for direct use.  Others simply haven't been documented yet.  
}
\usage{
residplot(model, formula, ...) 
}
\arguments{
  \item{model}{Undocumented}
  \item{formula}{Undocumented}
    \item{\dots}{ arguments to be passed to \code{fun} }
  }
\details{
 These functions are undocumented.  Some are internal and not intended
   for direct use.  Others simply haven't been documented yet.
   }
\author{Gregory R. Warnes \email{gregory.r.warnes@pfizer.com}}
\keyword{misc}
